extract webservice logic
